chore: refactor yatai_client and rest_api_client to cloud
